Crate aptos_gas

This crate is the core of the gas metering system of the Aptos blockchain.

More specifically, it

Is home to the gas meter implementation

Defines the gas parameters and formulae for instructions

Defines the gas parameters for transactions

Sets the initial values for all gas parameters, including the instruction, transaction move-stdlib and aptos-framework ones.

Defines a bi-directional mapping between the (Rust) gas parameter structs and their corresponding representation on-chain.

The reason why we need two different representations is that they serve different purposes:

The Rust structs are used for quick (static) lookups by the gas meter and native functions when calculating gas costs.

The on-chain gas schedule needs to be extensible and unordered so we can upgrate it easily in the future.

Modules

gen

Structs

AbstractValueSizeGasParameters

AptosGasMeter

The official gas meter used inside the Aptos VM. It maintains an internal gas counter, measured in internal gas units, and carries an environment consisting all the gas parameters, which it can lookup when performing gas calculations.

AptosGasParameters

Gas parameters for everything that is needed to run the Aptos blockchain, including instructions, transactions and native functions from various packages.

GasQuantity

An opaque representation of a certain quantity, with the unit being encoded in the type. This type implements checked addition and subtraction, and only permits type-safe multiplication.

InstructionGasParameters

MiscGasParameters

Miscellaneous gas parameters.

NativeGasParameters

Gas parameters for all native functions.

StorageGasParameters

TransactionGasParameters

UnitDiv

A derived unit resulted from the division of two given units. This is used to permit type-safe multiplications.

Enums

AbstractValueUnit

Unit of abstract value size – a conceptual measurement of the memory space a Move value occupies.

Arg

Unit for counting arguments.

Byte

Unit for counting bytes.

GasUnit

Unit of (external) gas.

InternalGasUnit

Unit of internal gas.

Octa

Unit of gas currency. 1 Octa = 10^-8 Aptos coins.

Constants

LATEST_GAS_FEATURE_VERSION

Traits

FromOnChainGasSchedule

A trait for converting from a map representation of the on-chain gas schedule.

InitialGasSchedule

A trait for defining an initial value to be used in the genesis.

ToOnChainGasSchedule

A trait for converting to a list of entries of the on-chain gas schedule.

Type Definitions

AbstractValueSize

AbstractValueSizePerArg

Fee

FeePerGasUnit

Gas

GasScalingFactor

InternalGas

InternalGasPerAbstractValueUnit

InternalGasPerArg

InternalGasPerByte

NumArgs

NumBytes