Trait aptos_gas::FromOnChainGasSchedule

pub trait FromOnChainGasSchedule: Sized {
    fn from_on_chain_gas_schedule(
        gas_schedule: &BTreeMap<String, u64>
    ) -> Option<Self>;
}

A trait for converting from a map representation of the on-chain gas schedule.

Required Methods

fn from_on_chain_gas_schedule(
gas_schedule: &BTreeMap<String, u64>
) -> Option<Self>

Constructs a value of this type from a map representation of the on-chain gas schedule. None should be returned when the gas schedule is missing some required entries. Unused entries should be safely ignored.