Trait aptos_gas::InitialGasSchedule

pub trait InitialGasSchedule: Sized {
    fn initial() -> Self;
}

A trait for defining an initial value to be used in the genesis.

Required Methods

fn initial() -> Self

Returns the initial value of this type, which is used in the genesis.